refactor(node): split into config, sequencer, and full crates
