+++ crates/node/config/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+edition.workspace = true
+license.workspace = true
+name = "katana-node-config"
+repository.workspace = true
+version.workspace = true
+
+[dependencies]
+http.workspace = true
+serde.workspace = true
+strum.workspace = true
+strum_macros.workspace = true
+thiserror.workspace = true
+
+[features]
+cartridge = []
+tee = []
+explorer = []
