+++ crates/node/config/src/lib.rs
@@ -1,0 +1,5 @@
+#![cfg_attr(not(test), warn(unused_crate_dependencies))]
+
+pub mod db;
+pub mod metrics;
+pub mod rpc;
