+++ crates/node/full/Cargo.toml
@@ -1,21 +1,18 @@
 [package]
 edition.workspace = true
 license.workspace = true
-name = "katana-node"
+name = "katana-full-node"
 repository.workspace = true
 version.workspace = true
 
 [dependencies]
+katana-node-config.workspace = true
 katana-chain-spec.workspace = true
-katana-core.workspace = true
 katana-db.workspace = true
 katana-executor.workspace = true
-katana-gateway-server.workspace = true
 katana-gateway-client.workspace = true
 katana-gateway-types.workspace = true
 katana-gas-price-oracle.workspace = true
-katana-grpc = { workspace = true, optional = true }
-katana-messaging.workspace = true
 katana-metrics.workspace = true
 katana-pipeline.workspace = true
 katana-pool.workspace = true
@@ -24,48 +21,23 @@ katana-primitives.workspace = true
 katana-provider.workspace = true
 katana-rpc-server.workspace = true
 katana-rpc-api.workspace = true
-katana-rpc-client.workspace = true
 katana-rpc-types.workspace = true
 katana-stage.workspace = true
 katana-tasks.workspace = true
-katana-tee = { workspace = true, optional = true }
 
 anyhow.workspace = true
 backon.workspace = true
-num-traits.workspace = true
 futures.workspace = true
 http.workspace = true
 jsonrpsee.workspace = true
-serde.workspace = true
 parking_lot.workspace = true
-thiserror.workspace = true
-tracing.workspace = true
-url.workspace = true
-
+serde.workspace = true
 strum.workspace = true
 strum_macros.workspace = true
-
+thiserror.workspace = true
+tracing.workspace = true
 tokio = { workspace = true, features = [ "time" ] }
 
 [features]
-paymaster = [
-	"katana-rpc-api/paymaster",
-	"katana-rpc-server/paymaster",
-]
-vrf = [
-	"katana-rpc-server/vrf",
-	"cartridge",
-]
-cartridge = [
-	"katana-rpc-api/cartridge",
-	"katana-rpc-server/cartridge",
-	"paymaster",
-]
-explorer = ["katana-rpc-server/explorer"]
-grpc = ["dep:katana-grpc"]
-native = ["katana-executor/native"]
-tee = ["dep:katana-tee", "katana-rpc-api/tee", "katana-rpc-server/tee"]
-tee-snp = ["tee", "katana-tee/snp"]
-
-[dev-dependencies]
-tempfile.workspace = true
+cartridge = ["katana-node-config/cartridge", "katana-rpc-api/cartridge", "katana-rpc-server/cartridge"]
+explorer = ["katana-node-config/explorer", "katana-rpc-server/explorer"]
