+++ crates/node/full/src/config.rs
@@ -1,0 +1,1 @@
+pub use katana_node_config::{db, metrics, rpc};
