+++ crates/node/sequencer/Cargo.toml
@@ -1,0 +1,69 @@
+[package]
+edition.workspace = true
+license.workspace = true
+name = "katana-sequencer-node"
+repository.workspace = true
+version.workspace = true
+
+[dependencies]
+katana-node-config.workspace = true
+katana-chain-spec.workspace = true
+katana-core.workspace = true
+katana-db.workspace = true
+katana-executor.workspace = true
+katana-gateway-server.workspace = true
+katana-gas-price-oracle.workspace = true
+katana-grpc = { workspace = true, optional = true }
+katana-messaging.workspace = true
+katana-metrics.workspace = true
+katana-pool.workspace = true
+katana-pool-api.workspace = true
+katana-primitives.workspace = true
+katana-provider.workspace = true
+katana-rpc-server.workspace = true
+katana-rpc-api.workspace = true
+katana-rpc-client.workspace = true
+katana-rpc-types.workspace = true
+katana-stage.workspace = true
+katana-tasks.workspace = true
+katana-tee = { workspace = true, optional = true }
+
+anyhow.workspace = true
+num-traits.workspace = true
+futures.workspace = true
+http.workspace = true
+jsonrpsee.workspace = true
+serde.workspace = true
+parking_lot.workspace = true
+thiserror.workspace = true
+tracing.workspace = true
+url.workspace = true
+
+strum.workspace = true
+strum_macros.workspace = true
+
+tokio = { workspace = true, features = [ "time" ] }
+
+[features]
+paymaster = [
+	"katana-rpc-api/paymaster",
+	"katana-rpc-server/paymaster",
+]
+vrf = [
+	"katana-rpc-server/vrf",
+	"cartridge",
+]
+cartridge = [
+	"katana-node-config/cartridge",
+	"katana-rpc-api/cartridge",
+	"katana-rpc-server/cartridge",
+	"paymaster",
+]
+explorer = ["katana-node-config/explorer", "katana-rpc-server/explorer"]
+grpc = ["dep:katana-grpc"]
+native = ["katana-executor/native"]
+tee = ["katana-node-config/tee", "dep:katana-tee", "katana-rpc-api/tee", "katana-rpc-server/tee"]
+tee-snp = ["tee", "katana-tee/snp"]
+
+[dev-dependencies]
+tempfile.workspace = true
