+++ crates/node/sequencer/src/config/mod.rs
@@ -1,12 +1,13 @@
 use std::sync::Arc;
 
-pub mod db;
+// Re-export shared config types
+pub use katana_node_config::{db, metrics, rpc};
+
+// Sequencer-specific config modules
 pub mod dev;
 pub mod execution;
 pub mod fork;
 pub mod gateway;
-pub mod metrics;
-pub mod rpc;
 pub mod sequencing;
 
 #[cfg(feature = "paymaster")]
