+++ crates/node/sequencer/src/lib.rs
@@ -1,7 +1,5 @@
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 
-pub mod full;
-
 pub mod config;
 pub mod exit;
