+++ tests/snos/src/main.rs
@@ -5,9 +5,9 @@ use cairo_vm::types::layout_name::LayoutName;
 use katana_chain_spec::rollup::{self, ChainConfigDir};
 use katana_chain_spec::ChainSpec;
 use katana_messaging::MessagingConfig;
-use katana_node::config::db::DbConfig;
-use katana_node::config::Config;
-use katana_node::{LaunchedNode, Node};
+use katana_sequencer_node::config::db::DbConfig;
+use katana_sequencer_node::config::Config;
+use katana_sequencer_node::{LaunchedNode, Node};
 use katana_primitives::block::BlockNumber;
 use katana_primitives::{address, Felt};
 use katana_provider::api::block::BlockNumberProvider;
